requirements.txt
accelerate==1.4.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
altair==5.4.1
annotated-types==0.7.0
anthropic==0.49.0
anyio==4.8.0
attrs==24.2.0
beautifulsoup4==4.12.3
blinker==1.8.2
CacheControl==0.14.0
cachetools==5.3.3
certifi==2024.12.14
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cohere==5.14.0
colorama==0.4.6
contourpy==1.3.1
cryptography==42.0.7
cycler==0.12.1
da_pkg_resources==0.0.1
dataclasses-json==0.6.7
Deprecated==1.2.15
dirtyjson==1.0.8
distlib==0.3.9
distro==1.9.0
eval_type_backport==0.2.2
fastavro==1.10.0
filelock==3.16.1
filetype==1.2.0
firebase==4.0.1
firebase-admin==6.5.0
Flask==3.0.3
Flask-Cors==4.0.1
fonttools==4.55.0
frozenlist==1.5.0
fsspec==2024.12.0
gcloud==0.18.3
gitdb==4.0.11
GitPython==3.1.43
google-api-core==2.19.0
google-api-python-client==2.129.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.16.0
google-cloud-storage==2.16.0
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.63.0
greenlet==3.0.3
griffe==1.6.0
groq==0.18.0
grpcio==1.63.0
grpcio-status==1.62.2
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.27.2
httpx-sse==0.4.0
huggingface-hub==0.27.1
idna==3.7
importlib_metadata==8.5.0
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
jiter==0.8.2
joblib==1.4.2
jsonpath-python==1.0.6
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jwcrypto==1.5.6
kiwisolver==1.4.7
llama-cloud==0.1.7
llama-index==0.12.10
llama-index-agent-openai==0.4.1
llama-index-cli==0.4.0
llama-index-core==0.12.10.post1
llama-index-embeddings-openai==0.3.1
llama-index-indices-managed-llama-cloud==0.6.3
llama-index-llms-openai==0.3.12
llama-index-multi-modal-llms-openai==0.4.2
llama-index-program-openai==0.3.1
llama-index-question-gen-openai==0.3.0
llama-index-readers-file==0.4.2
llama-index-readers-llama-parse==0.4.0
llama-parse==0.5.19
logfire-api==3.6.4
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.24.1
matplotlib==3.9.2
mdurl==0.1.2
mistralai==1.5.0
mpmath==1.3.0
msgpack==1.0.8
multidict==6.1.0
mypy-extensions==1.0.0
narwhals==1.8.4
nest-asyncio==1.6.0
networkx==3.4.2
nltk==3.9.1
numpy==2.1.1
oauth2client==4.1.3
ollama==0.4.5
openai==1.65.2
opencv-python==4.10.0.84
opentelemetry-api==1.30.0
outcome==1.3.0.post0
packaging==24.1
pandas==2.2.3
pillow==10.4.0
platformdirs==4.3.6
playwright==1.47.0
playwright-recaptcha==0.5.1
pluggy==1.5.0
propcache==0.2.1
proto-plus==1.23.0
protobuf==4.25.3
psutil==7.0.0
pyarrow==17.0.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pycryptodome==3.20.0
pydantic==2.10.4
pydantic-ai==0.0.31
pydantic-ai-slim==0.0.31
pydantic-graph==0.0.31
pydantic_core==2.27.2
pydeck==0.9.1
pydub==0.25.1
pyee==12.0.0
Pygments==2.18.0
PyJWT==2.8.0
pyparsing==3.1.2
pypdf==5.1.0
Pyrebase4==4.8.0
PySocks==1.7.1
pytest==8.3.3
pytest-base-url==2.1.0
pytest-playwright==0.5.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jwt==4.1.0
python-slugify==8.0.4
pytz==2024.2
PyYAML==6.0.2
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
requests-toolbelt==0.10.1
rich==13.8.1
rpds-py==0.20.0
rsa==4.9
safetensors==0.5.1
scikit-learn==1.6.1
scipy==1.15.1
seaborn==0.13.2
selenium==4.25.0
setuptools==70.0.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
SpeechRecognition==3.10.4
SQLAlchemy==2.0.36
streamlit==1.38.0
striprtf==0.0.26
sympy==1.13.1
tenacity==8.3.0
text-unidecode==1.3
threadpoolctl==3.5.0
tiktoken==0.8.0
tokenizers==0.21.0
toml==0.10.2
torch==2.5.1
tornado==6.4.1
tqdm==4.67.0
transformers==4.47.1
trio==0.26.2
trio-websocket==0.11.1
types-requests==2.32.0.20250301
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.3.0
virtualenv==20.29.3
watchdog==4.0.2
websocket-client==1.8.0
Werkzeug==3.0.3
wrapt==1.17.0
wsproto==1.2.0
yarl==1.18.3
zipp==3.21.0
